+++ CHANGELOG.md
@@ -30,10 +30,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Vulnerability: API Key Scanner
 - Vulnerability: Database Connection String
 - Vulnerability: MySQL Username Disclosure
+- Vulnerability: 403 Bypass
 - Vulnerability: Firebase Database Unauthorized Access
 - GIF Favicon
 - Unauthenticated Gitlab SSRF CVE 2021-22214 Demonstration
 - Software Versions List
+- Error Handling
 
 ### Changed
 - Restructure Project